Private data hidden, express middleware added...
I added express middleware (authentication) to the application. I also hid credentials (password and the tokens array) so that they don't get sent to the client after any request. In addition, I made a '/users/signout' endpoint and a '/users/signout/all' endpoint. Sign out deletes the current token that's used in the current session and sign out all signs the user out from all of the sessions (or devices).